requirements.txt
flask
python-dotenv
aiohttp
requests
google-generativeai
Flask-SQLAlchemy
psycopg2-binary